+++ project/src/components/login-form/login-form.tsx
@@ -1,19 +1,14 @@
-import { ChangeEvent, FormEvent, useState } from 'react';
 import cn from 'classnames';
 
 import Spinner from '../spinner/spinner';
 
-import { useAppDispatch, useAppSelector } from '../../hooks';
-
-import { loginAction } from '../../store/api-actions';
+import { useAppSelector } from '../../hooks';
 
 import { State } from '../../types/state';
 import { FetchStatus } from '../../utils/const';
 
 import styles from './login-form.module.css';
-
-const REG_EXP_EMAIL = /^\S+@[aA-zZ]{2,10}\.[aA-zZ]{2,3}$/;
-const REG_EXP_PASSWORD = /([0-9]{1}[aA-zZ]{1})|([aA-zZ]{1}[0-9]{1})/i;
+import useLoginForm from '../../hooks/use-login-form';
 
 const fields = {
   email: {
@@ -26,59 +21,13 @@ const fields = {
   },
 };
 
-interface LoginField {
-  value: string;
-  regexp: RegExp;
-  error: boolean;
-  errorText: string;
-}
-
-type InitialState = { [key: string]: LoginField };
-
 function LoginForm(): JSX.Element {
-  const dispatch = useAppDispatch();
   const { loginStatus } = useAppSelector((state: State) => state);
 
-  const [formState, setFormState] = useState<InitialState>({
-    email: {
-      value: '',
-      regexp: REG_EXP_EMAIL,
-      error: false,
-      errorText: 'Email is not entered correctly',
-    },
-    password: {
-      value: '',
-      regexp: REG_EXP_PASSWORD,
-      error: false,
-      errorText: 'Enter at least 1 number and 1 letter',
-    },
-  });
+  const { formState, handleChange, handleSubmit } = useLoginForm();
 
   const isPending = loginStatus === FetchStatus.PENDING;
   const isValid = Object.values(formState).some(({ error }) => error === true);
-
-  const handleChange = (evt: ChangeEvent<HTMLInputElement>) => {
-    const { value, name } = evt.target;
-
-    const regExp = formState[name].regexp;
-    const isValid = regExp.test(value);
-
-    setFormState((prevState) => ({
-      ...prevState,
-      [name]: {
-        ...prevState[name],
-        error: !isValid,
-        value,
-      },
-    }));
-  };
-
-  const handleSubmit = (evt: FormEvent) => {
-    evt.preventDefault();
-    const authData = { email: formState.email.value, password: formState.password.value };
-
-    dispatch(loginAction(authData));
-  };
 
   return (
     <section className="login">
